Move `Switch` out of single-file directory
